+++ src/gpu/soft/zn.cc
@@ -26,34 +26,36 @@
 
 #define _IN_ZN
 
+#include <stdint.h>
+
 #include "gpu/soft/externals.h"
 
 // --------------------------------------------------- //
 // - psx gpu plugin interface prototypes-------------- //
 // --------------------------------------------------- //
 
-// long GPUopen(HWND hwndGPU);
+// int32_t GPUopen(HWND hwndGPU);
 void GPUdisplayText(char* pText);
-void GPUdisplayFlags(unsigned long dwFlags);
+void GPUdisplayFlags(uint32_t dwFlags);
 void GPUmakeSnapshot(void);
-long GPUinit();
-long GPUclose();
-long GPUshutdown();
+int32_t GPUinit();
+int32_t GPUclose();
+int32_t GPUshutdown();
 void GPUcursor(int iPlayer, int x, int y);
 void GPUupdateLace(void);
-unsigned long GPUreadStatus(void);
-void GPUwriteStatus(unsigned long gdata);
-void GPUreadDataMem(unsigned long* pMem, int iSize);
-unsigned long GPUreadData(void);
-void GPUwriteDataMem(unsigned long* pMem, int iSize);
-void GPUwriteData(unsigned long gdata);
-void GPUsetMode(unsigned long gdata);
-long GPUgetMode(void);
-long GPUdmaChain(unsigned long* baseAddrL, unsigned long addr);
-long GPUconfigure(void);
+uint32_t GPUreadStatus(void);
+void GPUwriteStatus(uint32_t gdata);
+void GPUreadDataMem(uint32_t* pMem, int iSize);
+uint32_t GPUreadData(void);
+void GPUwriteDataMem(uint32_t* pMem, int iSize);
+void GPUwriteData(uint32_t gdata);
+void GPUsetMode(uint32_t gdata);
+int32_t GPUgetMode(void);
+int32_t GPUdmaChain(uint32_t* baseAddrL, uint32_t addr);
+int32_t GPUconfigure(void);
 void GPUabout(void);
-long GPUtest(void);
-long GPUfreeze(unsigned long ulGetFreezeData, void* pF);
+int32_t GPUtest(void);
+int32_t GPUfreeze(uint32_t ulGetFreezeData, void* pF);
 void GPUgetScreenPic(unsigned char* pMem);
 void GPUshowScreenPic(unsigned char* pMem);
 #ifndef _WIN32
@@ -64,7 +66,7 @@ void GPUkeypressed(int keycode);
 // - zn gpu interface -------------------------------- //
 // --------------------------------------------------- //
 
-unsigned long dwGPUVersion = 0;
+uint32_t dwGPUVersion = 0;
 int iGPUHeight = 512;
 int iGPUHeightMask = 511;
 int GlobalTextIL = 0;
@@ -75,10 +77,10 @@ int iTileCheat = 0;
 // --------------------------------------------------- //
 
 typedef struct GPUOTAG {
-    unsigned long Version;         // Version of structure - currently 1
-    long hWnd;                     // Window handle
-    unsigned long ScreenRotation;  // 0 = 0CW, 1 = 90CW, 2 = 180CW, 3 = 270CW = 90CCW
-    unsigned long GPUVersion;      // 0 = a, 1 = b, 2 = c
+    uint32_t Version;         // Version of structure - currently 1
+    int32_t hWnd;                     // Window handle
+    uint32_t ScreenRotation;  // 0 = 0CW, 1 = 90CW, 2 = 180CW, 3 = 270CW = 90CCW
+    uint32_t GPUVersion;      // 0 = a, 1 = b, 2 = c
     const char* GameName;          // NULL terminated string
     const char* CfgFile;           // NULL terminated string
 } GPUConfiguration_t;
